+++ src/main/java/tigerworkshop/webapphardwarebridge/Constants.java
@@ -3,5 +3,5 @@ package tigerworkshop.webapphardwarebridge;
 public class Constants {
     public static final String APP_NAME = "茶坊村本地打印助手";
     public static final String APP_ID = "zodance.printhelper";
-    public static final String VERSION = "1.2.0";
+    public static final String VERSION = "1.3.0";
 }