+++ src/main/java/tigerworkshop/webapphardwarebridge/dtos/LocalPrintDeviceActivationDTO.java
@@ -1,0 +1,9 @@
+package tigerworkshop.webapphardwarebridge.dtos;
+
+import lombok.Data;
+
+@Data
+public class LocalPrintDeviceActivationDTO {
+    private String shopId;
+    private Long activationTaskId;
+}