+++ src/main/java/tigerworkshop/webapphardwarebridge/dtos/LocalPrintDeviceContextDTO.java
@@ -1,0 +1,14 @@
+package tigerworkshop.webapphardwarebridge.dtos;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class LocalPrintDeviceContextDTO {
+    private String deviceId;
+    private String shopId;
+    private Long activationTaskId;
+}